@@ -1,9 +1,9 @@
 [project]
 name = "ai-orchestra-gateway"
-version = "0.3.0"
+version = "0.6.1"
 description = "AI Legal Ops Gateway"
 authors = [{name = "AI Legal Ops Team"}]
-requires-python = ">=3.11"
+requires-python = ">=3.12"
 
 [tool.ruff]
 line-length = 88
