@@ -1,6 +1,6 @@
 import logging
-from typing import Any
 from app.services.privacy import DataPrivacyShield
+from app.core.middleware import get_request_id
 
 
 class PrivacyLogFilter(logging.Filter):
@@ -12,19 +12,12 @@ class PrivacyLogFilter(logging.Filter):
     def filter(self, record: logging.LogRecord) -> bool:
         """
         Filter log record by sanitizing PII from message and arguments.
-        
-        Args:
-            record: LogRecord to filter
-            
-        Returns:
-            True (always - we modify but don't block records)
         """
         # Sanitize the main log message
         if isinstance(record.msg, str):
             record.msg, _ = DataPrivacyShield.sanitize(record.msg)
         
-        # Handle arguments if they are strings
-        # Note: This handles cases like logger.info("User %s logged in", user_email)
+        # Handle arguments
         if record.args:
             new_args = []
             for arg in record.args:
@@ -35,4 +28,16 @@ class PrivacyLogFilter(logging.Filter):
                     new_args.append(arg)
             record.args = tuple(new_args)
 
+        return True
+
+
+class RequestIdFilter(logging.Filter):
+    """
+    Logging filter that adds the current request ID to log records.
+    Ensures request_id is always present to prevent KeyError in log formatting.
+    """
+    def filter(self, record: logging.LogRecord) -> bool:
+        # Always set request_id, fallback to "system" if not in request context
+        if not hasattr(record, 'request_id') or not record.request_id:
+            record.request_id = get_request_id() or "system"
         return True
